feat: refactor feature resolution strategies to use EvaluateAsync method for improved multi-provider support
